Move SuicideSystem into its own directory
