+++ Content.Client/Suicide/SuicideSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Suicide;
+
+namespace Content.Client.Suicide;
+
+public sealed class SuicideSystem : SharedSuicideSystem;
