+++ Content.Shared/Execution/SharedExecutionSystem.cs
@@ -1,17 +1,17 @@
 using Content.Shared.ActionBlocker;
-using Content.Shared.Chat;
 using Content.Shared.CombatMode;
 using Content.Shared.Damage.Components;
 using Content.Shared.Database;
 using Content.Shared.DoAfter;
 using Content.Shared.IdentityManagement;
+using Content.Shared.Interaction.Events;
 using Content.Shared.Mobs.Components;
 using Content.Shared.Mobs.Systems;
 using Content.Shared.Popups;
+using Content.Shared.Suicide;
 using Content.Shared.Verbs;
 using Content.Shared.Weapons.Melee;
 using Content.Shared.Weapons.Melee.Events;
-using Content.Shared.Interaction.Events;
 using Robust.Shared.Audio.Systems;
 
 namespace Content.Shared.Execution;
